Upgrade Kusto SDK
Kusto SDK released a new major version in november 2022 with some breaking changes:
- The minimum JDK version for the SDK is now up to jdk 11
- Move to Blob Storage v12
- Remove org.json library to support secure systems that don't allow it. Some of the exception signatures have been changed to custom exception.

The latter change is the one having the most impacts on the connector: org.json is removed in favor of Jackson.
The KustoSinkConfig and KustoSinkTask were updated to take advantage of Jackson capabilities, in particular the mapping configuration is now deserialized to a POJO rather than doing operations on raw JSON objects.